+++ streamingcli/jupyter/notebook_converter.py
@@ -1,10 +1,11 @@
 import argparse
 import json
 import os
+import re
 import shlex
 import sys
 from dataclasses import asdict, dataclass, field
-from typing import Any, Dict, List, Sequence
+from typing import Any, Dict, List, Optional, Sequence, Tuple, cast
 
 import autopep8
 import nbformat
@@ -70,6 +71,13 @@ class NotebookConverter:
 
     notebook_path: str
     """String representing path to .ipynb file"""
+
+    secrets_paths: Dict[str, str]
+    """Mapping from secret variable name to path of the file holding the secret."""
+
+    hidden_variable_counter: int
+    """Counter of hidden variables that are read from ENVs"""
+
     _udf_parser = argparse.ArgumentParser()
     """Argument parser for %register_udf magic"""
     _udf_parser.add_argument("--function_name", "-n")
@@ -77,15 +85,24 @@ class NotebookConverter:
     _udf_parser.add_argument("--language", "-l")
     _udf_parser.add_argument("--remote_path", "-r")
     _udf_parser.add_argument("--local_path", "-p")
+
     _load_config_parser = argparse.ArgumentParser()
     """Argument parser for %load_config magic"""
     _load_config_parser.add_argument("--path", "-p")
+
+    _load_secret_parser = argparse.ArgumentParser()
+    """Argument parser for %load_secret magic"""
+    _load_secret_parser.add_argument("--path", "-p")
+    _load_secret_parser.add_argument("varname")
+
     _flink_execute_sql_file_parser = argparse.ArgumentParser()
     """Argument parser for %flink_execute_sql_file magic"""
     _flink_execute_sql_file_parser.add_argument("--path", "-p")
 
-    def __init__(self, notebook_path: str):
+    def __init__(self, notebook_path: str, secrets_paths: Optional[Dict[str, str]]):
         self.notebook_path = notebook_path
+        self.secrets_paths = secrets_paths or {}
+        self.hidden_variable_counter = 0
 
     """
     Read and convert Jupyter Notebook
@@ -99,10 +116,10 @@ class NotebookConverter:
             notebook = self._load_notebook(self.notebook_path)
             code_cells = filter(lambda _: _.cell_type == "code", notebook.cells)
             script_entries: List[NotebookEntry] = []
-            init_entries = NotebookConverter._read_init_sql(
-                os.path.dirname(self.notebook_path)
+            script_entries.extend(self._load_secrets_from_scli_config())
+            script_entries.extend(
+                self._read_init_sql(os.path.dirname(self.notebook_path))
             )
-            script_entries.extend(init_entries)
             for cell in code_cells:
                 entries = self._get_notebook_entry(
                     cell, os.path.dirname(self.notebook_path)
@@ -123,29 +140,27 @@ class NotebookConverter:
         with open(notebook_path, "r+") as notebook_file:
             return nbformat.reads(notebook_file.read(), as_version=4)
 
-    @staticmethod
     def _get_notebook_entry(
-        cell: nbformat.NotebookNode, notebook_dir: str
+        self, cell: nbformat.NotebookNode, notebook_dir: str
     ) -> Sequence[NotebookEntry]:
         if cell.source.startswith("%"):
-            return NotebookConverter._handle_magic_cell(cell, notebook_dir)
+            return self._handle_magic_cell(cell, notebook_dir)
         elif not cell.source.startswith("##"):
             return [Code(value=cell.source)]
         else:
             return []
 
-    @staticmethod
     def _handle_magic_cell(
-        cell: nbformat.NotebookNode, notebook_dir: str
+        self, cell: nbformat.NotebookNode, notebook_dir: str
     ) -> Sequence[NotebookEntry]:
         source = cell.source
         if source.startswith("%%flink_execute_sql"):
             sql_statement = "\n".join(source.split("\n")[1:])
             if NotebookConverter._skip_statement(sql_statement):
                 return []
-            return [Sql(value=sql_statement)]
+            return self._convert_sql_statement_to_python_instructions(sql_statement)
         if source.startswith("%flink_execute_sql_file"):
-            return NotebookConverter._get_statements_from_file(source, notebook_dir)
+            return self._get_statements_from_file(source, notebook_dir)
         if source.startswith("%flink_register_function"):
             args = NotebookConverter._udf_parser.parse_args(shlex.split(source)[1:])
             return [
@@ -167,6 +182,14 @@ class NotebookConverter:
                 if args.remote_path is not None
                 else RegisterLocalJar(local_path=args.local_path)
             ]
+        if source.startswith("%load_secret_file"):
+            args = NotebookConverter._load_secret_parser.parse_args(
+                shlex.split(source)[1:]
+            )
+            file_path = (
+                args.path if os.path.isabs(args.path) else f"{notebook_dir}/{args.path}"
+            )
+            return [NotebookConverter._load_secret_from_file(args.varname, file_path)]
         return []
 
     @staticmethod
@@ -185,9 +208,8 @@ class NotebookConverter:
         all_variable_strings = "\n".join(variable_strings)
         return Code(value=f"{all_variable_strings}")
 
-    @staticmethod
     def _get_statements_from_file(
-        source: str, notebook_dir: str
+        self, source: str, notebook_dir: str
     ) -> Sequence[NotebookEntry]:
         args = NotebookConverter._flink_execute_sql_file_parser.parse_args(
             shlex.split(source)[1:]
@@ -196,15 +218,26 @@ class NotebookConverter:
             args.path if os.path.isabs(args.path) else f"{notebook_dir}/{args.path}"
         )
         with open(file_path, "r") as f:
-            statements = [Sql(value=s.rstrip(";")) for s in sqlparse.split(f.read())]
-        return statements
+            instructions = [
+                instruction
+                for statement in sqlparse.split(f.read())
+                for instruction in self._convert_sql_statement_to_python_instructions(
+                    statement.rstrip(";")
+                )
+            ]
+        return instructions
 
-    @staticmethod
-    def _read_init_sql(notebook_dir: str) -> Sequence[NotebookEntry]:
+    def _read_init_sql(self, notebook_dir: str) -> Sequence[NotebookEntry]:
         file_path = f"{notebook_dir}/init.sql"
         if os.path.exists(file_path):
             with open(file_path, "r") as f:
-                return [Sql(value=s.rstrip(";")) for s in sqlparse.split(f.read())]
+                return [
+                    instruction
+                    for statement in sqlparse.split(f.read())
+                    for instruction in self._convert_sql_statement_to_python_instructions(
+                        statement.rstrip(";")
+                    )
+                ]
         return []
 
     @staticmethod
@@ -246,6 +279,52 @@ class NotebookConverter:
             statement.strip().lower().startswith(("select", "show", "desc", "explain"))
         )
 
+    def _convert_sql_statement_to_python_instructions(
+        self, sql_statement: str
+    ) -> List[NotebookEntry]:
+        sql_statement, env_var_loads = self._convert_hidden_variables_to_env_vars(
+            sql_statement
+        )
+        return env_var_loads + [Sql(value=sql_statement)]
 
-def convert_notebook(notebook_path: str) -> ConvertedNotebook:
-    return NotebookConverter(notebook_path).convert_notebook()
+    def _convert_hidden_variables_to_env_vars(
+        self, sql_statement: str
+    ) -> Tuple[str, List[NotebookEntry]]:
+        hidden_var_pattern = r"(\$\{[_a-zA-Z][_a-zA-Z0-9]*\})"
+        hidden_list = cast(List[str], re.findall(hidden_var_pattern, sql_statement))
+        hidden_env_load_instructions: List[NotebookEntry] = (
+            [Code(value="import os")] if len(hidden_list) > 0 else []
+        )
+        for hidden in hidden_list:
+            var_name = hidden[2:-1].strip()
+
+            hidden_variable_index = self.hidden_variable_counter
+            self.hidden_variable_counter = self.hidden_variable_counter + 1
+
+            hidden_env_variable = f"__env_var_{hidden_variable_index}__{var_name}"
+            hidden_env_load_instructions.append(
+                Code(value=f'{hidden_env_variable} = os.environ["{var_name}"]')
+            )
+
+            sql_statement = sql_statement.replace(
+                hidden, "{" + hidden_env_variable + "}"
+            )
+        return sql_statement, hidden_env_load_instructions
+
+    def _load_secrets_from_scli_config(self) -> Sequence[NotebookEntry]:
+        return [
+            self._load_secret_from_file(var_name, filepath)
+            for var_name, filepath in self.secrets_paths.items()
+        ]
+
+    @staticmethod
+    def _load_secret_from_file(var_name: str, filepath: str) -> NotebookEntry:
+        return Code(
+            value=f'with open("{filepath}", "r") as secret_file:\n    {var_name} = secret_file.read().rstrip()'
+        )
+
+
+def convert_notebook(
+    notebook_path: str, secrets_paths: Optional[Dict[str, str]] = None
+) -> ConvertedNotebook:
+    return NotebookConverter(notebook_path, secrets_paths).convert_notebook()